Increase parity between "pitch" & "normal" loaders
Webpack's documentation describes the default behavior of "pitching"
loaders as follows [1]:

> [...] If a loader delivers a result in the pitch method the process
> turns around and skips the remaining loaders, continuing with the
> calls to the more left loaders. [...]

This implicitly suggests that the "pitching" behavior will not take
place when a value is not returned. The loader will be "skipped", and
traversal through the chain of loaders will continue.

In synchronous "pitching" loaders, the skipping behavior can be
triggered by explicitly returning the `undefined` value or by omitting a
return statement altogether.

In asynchronous "pitching" loaders which signal completion via a
callback function, the skipping behavior can only be triggered by
invoking the callback function with zero or 1 arguments. If a second
argument value is specified as `undefined`, that value is interpreted as
invalid source code (rather than a signal to continue traversing the
chain of loaders). This argument-length-driven behavior is inconvenient
because it requires consumers maintain a dedicated code path for the
described condition (which may explain why such APIs are relatively rare
in idiomatic Node.js code).

In asynchronous "pitching" loaders which signal completion via a Promise
value, this same implementation detail makes it impossible to trigger
the skipping behavior.

Consistently interpret the value `undefined` as a signal to skip the
loader, whether specified as return value from synchronous loaders, an
asynchronous callback value, or an asynchronous Promise resolution
value.

[1] https://webpack.js.org/api/loaders/#pitching-loader